requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
streamlit==1.28.0
chromadb==0.4.15
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
httpx==0.25.0
pandas==2.1.1
python-dotenv==1.0.0
langchain==0.0.312
# deepseek-ai==0.1.1  # Commented out as not available 